add owners list
