+++ src/App.jsx
@@ -1,63 +1,11 @@
 import { Routing } from "./router/Routing.jsx";
 
-const owners = [
-  {
-    id: 1,
-    tipodedocumento: "CC",
-    identificacion: "1117800000",
-    nombre: "Andrés López Ortega",
-    email: "andres@gmail.com",
-    direccion: "Barrio San Luis",
-    telefono: "3209000000",
-  },
-
-  {
-    id: 1,
-    tipodedocumento: "CC",
-    identificacion: "1117800000",
-    nombre: "Andrés López Ortega",
-    email: "andres@gmail.com",
-    direccion: "Barrio San Luis",
-    telefono: "3209000000",
-  },
-
-  {
-    id: 1,
-    tipodedocumento: "CC",
-    identificacion: "1117800000",
-    nombre: "Andrés López Ortega",
-    email: "andres@gmail.com",
-    direccion: "Barrio San Luis",
-    telefono: "3209000000",
-  },
-
-  {
-    id: 1,
-    tipodedocumento: "CC",
-    identificacion: "1117800000",
-    nombre: "Andrés López Ortega",
-    email: "andres@gmail.com",
-    direccion: "Barrio San Luis",
-    telefono: "3209000000",
-  },
-
-  {
-    id: 1,
-    tipodedocumento: "CC",
-    identificacion: "1117800000",
-    nombre: "Andrés López Ortega",
-    email: "andres@gmail.com",
-    direccion: "Barrio San Luis",
-    telefono: "3209000000",
-  },
-];
-
-function App() {
+const App = () => {
   return (
     <div className="App">
       <Routing></Routing>
     </div>
   );
-}
+};
 
 export default App;
