+++ src/Components/dog/CreateDog.jsx
@@ -1,5 +1,5 @@
-function CreateDog() {
+const CreateDog = () => {
   return <div>hola mundo</div>;
-}
+};
 
 export default CreateDog;
