+++ src/assets/css/login.css
@@ -1,7 +1,7 @@
 .login-page {
+  background: no-repeat url(../images/forma-uno.png) bottom right;
   & .logo {
     max-width: 150px;
-    margin: 20px;
   }
   & .dog-walker {
     max-width: 90%;
